95 Create dialog with ask before delete
Problem: When user tab on delete button, app not asking bout
confirmation.

Solution: Has been added dialog with asking before delete

93 Check if title todo is exist on device

Solution: Add method checking for names file. if title equals new title,
todo exist and can't save this. User should rename new todo.

- Create view with bar for search for SearchActivity
- If user search title and text from input match with titles in db has been displayed in list.

101 Remove whitespace on last index.

Problem: If user press space on last index of string app will save it as "example_.db". That is bug, becase " " are replace to "_" and on view this is show as this same todos!

Solurion: Written method for checking last char of string and if it is " " it is delete.

update icon of none reminder in todo preview