101 Remove whitespace on last index.
Problem: If user press space on last index of string app will save it as "example_.db". That is bug, becase " " are replace to "_" and on view this is show as this same todos!

Solurion: Written method for checking last char of string and if it is " " it is delete.